+++ ark_auth/src/server/route/auth/provider/microsoft.rs
@@ -1,15 +1,13 @@
-use crate::{
-    core,
-    server::{
-        route::auth::oauth2::{oauth2_redirect, CallbackQuery, UrlResponse},
-        route_response_json, ConfigurationOauth2Provider, Data, Error, FromJsonValue, Oauth2Error,
-    },
+use crate::core;
+use crate::server::route::auth::provider::{
+    oauth2_redirect, Oauth2CallbackQuery, Oauth2UrlResponse,
 };
-use actix_web::{
-    http::{header, StatusCode},
-    middleware::identity::Identity,
-    web, HttpResponse, ResponseError,
+use crate::server::{
+    route_response_json, ConfigurationProviderOauth2, Data, Error, FromJsonValue, Oauth2Error,
 };
+use actix_web::http::{header, StatusCode};
+use actix_web::middleware::identity::Identity;
+use actix_web::{web, HttpResponse, ResponseError};
 use futures::{future, Future};
 use oauth2::prelude::*;
 use oauth2::{
@@ -23,7 +21,7 @@ struct MicrosoftUser {
     mail: String,
 }
 
-pub fn request_handler(
+fn oauth2_request_handler(
     data: web::Data<Data>,
     id: Identity,
 ) -> impl Future<Item = HttpResponse, Error = actix_web::Error> {
@@ -34,18 +32,18 @@ pub fn request_handler(
         .then(route_response_json)
 }
 
-fn request_inner(data: &Data, id: Option<String>) -> Result<UrlResponse, Error> {
+fn request_inner(data: &Data, id: Option<String>) -> Result<Oauth2UrlResponse, Error> {
     core::key::authenticate_service(data.driver(), id)
         .map_err(Into::into)
         .and_then(|service| microsoft_authorise(&data, &service).map_err(Into::into))
-        .map(|url| UrlResponse { url })
+        .map(|url| Oauth2UrlResponse { url })
 }
 
-pub fn callback_handler(
+fn oauth2_callback_handler(
     data: web::Data<Data>,
     query: web::Query<serde_json::Value>,
 ) -> impl Future<Item = HttpResponse, Error = actix_web::Error> {
-    CallbackQuery::from_value(query.into_inner())
+    Oauth2CallbackQuery::from_value(query.into_inner())
         .and_then(|query| {
             web::block(move || {
                 let (service_id, access_token) =
@@ -66,7 +64,7 @@ pub fn callback_handler(
                     data.driver(),
                     service_id,
                     &email,
-                    data.configuration().token_exp(),
+                    data.configuration().token_expiration_time(),
                 )
                 .map_err(Into::into)
             })
@@ -84,7 +82,7 @@ fn microsoft_authorise(data: &Data, service: &core::Service) -> Result<String, E
     let code_verifier = PkceCodeVerifierS256::new_random();
 
     // Generate the authorisation URL to redirect.
-    let client = microsoft_client(data.configuration().oauth2_microsoft())?;
+    let client = microsoft_client(data.configuration().provider_microsoft_oauth2())?;
     let (authorize_url, state) = client.authorize_url_extension(
         &ResponseType::new("code".to_string()),
         CsrfToken::new_random,
@@ -113,7 +111,7 @@ fn microsoft_callback(data: &Data, code: &str, state: &str) -> Result<(i64, Stri
     let params: Vec<(&str, &str)> = vec![("code_verifier", &csrf.value)];
 
     // Exchange the code with a token.
-    let client = microsoft_client(data.configuration().oauth2_microsoft())?;
+    let client = microsoft_client(data.configuration().provider_microsoft_oauth2())?;
     let code = AuthorizationCode::new(code.to_owned());
     let token = client
         .exchange_code_extension(code, &params)
@@ -150,7 +148,7 @@ fn microsoft_api_user_email(
         .map(|res| res.mail)
 }
 
-fn microsoft_client(provider: Option<&ConfigurationOauth2Provider>) -> Result<BasicClient, Error> {
+fn microsoft_client(provider: Option<&ConfigurationProviderOauth2>) -> Result<BasicClient, Error> {
     let provider = provider.ok_or(Error::Oauth2(Oauth2Error::Disabled))?;
 
     let graph_client_id = ClientId::new(provider.client_id.to_owned());
@@ -176,4 +174,13 @@ fn microsoft_client(provider: Option<&ConfigurationOauth2Provider>) -> Result<Ba
     ))
     .set_auth_type(AuthType::RequestBody)
     .set_redirect_url(RedirectUrl::new(redirect_url)))
+}
+
+/// Version 1 API authentication oauth2 scope.
+pub fn api_v1_scope() -> actix_web::Scope {
+    web::scope("/microsoft").service(
+        web::resource("/oauth2")
+            .route(web::post().to_async(oauth2_request_handler))
+            .route(web::get().to_async(oauth2_callback_handler)),
+    )
 }